Fixes after replace of `int` with `size_t` in `Bottle`.

+++ src/libraries/iKin/include/iCub/iKin/iKinSlv.h
@@ -269,7 +269,7 @@ protected:
 
     bool contMode;
     bool isNew;
-    int  maxLen;
+    size_t  maxLen;
     int  pose;
 
     double  token;
